packages/monitor-engine/requirements.txt
# OCME bd Monitor Engine — Dependências Python
# Instale com: pip install -r requirements.txt

# Blockchain
web3>=6.0.0,<7.0.0

# Telegram Bot
pyTelegramBotAPI>=4.14.0,<5.0.0

# HTTP
requests>=2.31.0,<3.0.0

# Configuração
python-dotenv>=1.0.0,<2.0.0

# Gráficos (dashboard)
matplotlib>=3.7.0,<4.0.0

# Utilidades (já no stdlib, mas explícito para clareza)
# sqlite3 — builtin
# threading — builtin
# argparse — builtin
# json — builtin
# logging — builtin
